Clarified instructions for step 6 of JavaScript exercise
Prior instructions for step 6) were unclear - "Executing your custom command from (5)" could be misunderstood as executing the function from the 5th index, which does not exist in the lecture provided. This would raise an undefined value. Clarified this as "step 5)".

Further clarification provided to output the result of the first identifier(word used in the prior step) of the goals array as this is what was requested in the exercise.